APLICACIONES _ AZURE/DelNegocio/ms-medicina/src/test/java/com/grupo1/infraestructure/adapter/TriajeAdapterTest.java
package com.grupo1.infraestructure.adapter;

import com.grupo1.domain.aggregates.dto.TriajeDTO;
import com.grupo1.domain.aggregates.request.RequestTriaje;
import com.grupo1.domain.aggregates.response.ResponseBase;
import com.grupo1.infraestructure.entity.ConsultaMedicaEntity;
import com.grupo1.infraestructure.entity.TriajeEntity;
import com.grupo1.infraestructure.mapper.GenericMapper;
import com.grupo1.infraestructure.repository.ConsultaMedicaRepository;
import com.grupo1.infraestructure.repository.TriajeRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.Arrays;
import java.util.HashSet;
import java.util.Optional;
import java.util.Set;

import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.*;

class TriajeAdapterTest {

    @Mock
    private TriajeRepository triajeRepository;
    @Mock
    private ConsultaMedicaRepository consultaMedicaRepository;
    @Mock
    private GenericMapper genericMapper;
    @InjectMocks
    private TriajeAdapter triajeAdapter;

    @BeforeEach
    void setUp(){
        MockitoAnnotations.initMocks(this);
    }

    @Test
    void buscarTriajeEntityOut_triageExist_returnsResponseBase() {
        // Arrange
        TriajeEntity triajeEntity = new TriajeEntity();
        Optional<TriajeEntity> optional = Optional.of(triajeEntity);
        Long id = 1L;

        when(triajeRepository.findById(id)).thenReturn(optional);

        // Act
        ResponseBase result = triajeAdapter.BuscarTriajeEntityOut(id);

        // Assert
        assertEquals(200, result.getCode());
        assertEquals("Registro encontrado con exito", result.getMessage());
        assertEquals(optional, result.getData());

        // Verify repository method call
        verify(triajeRepository).findById(id);
    }

    @Test
    void buscarTriajeEntityOut_triageNotExist_returnsResponseBase() {
        // Arrange
        when(triajeRepository.findById(1L)).thenReturn(Optional.empty());

        // Act
        ResponseBase result = triajeAdapter.BuscarTriajeEntityOut(1L);

        // Assert
        assertEquals(404, result.getCode());
        assertEquals("Registro no encontrado", result.getMessage());
        assertEquals(null, result.getData());

        // Verify repository method call
        verify(triajeRepository).findById(1L);
    }

    @Test
    void buscarTriajeDtoOut_triageExist_returnsResponseBase() {
        // Arrange
        TriajeEntity triajeEntity = new TriajeEntity();
        triajeEntity.setIdTriaje(1L);

        TriajeDTO triajeDTO = new TriajeDTO();
        triajeDTO.setIdTriaje(1L);

        when(triajeRepository.findById(1L)).thenReturn(Optional.of(triajeEntity));
        when(genericMapper.mapTriajeEntityToTriajeDTO(triajeEntity)).thenReturn(triajeDTO);

        // Act
        ResponseBase result = triajeAdapter.BuscarTriajeDtoOut(1L);

        // Assert
        assertEquals(200, result.getCode());
        assertEquals("Registro encontrado con exito", result.getMessage());
        assertEquals(triajeDTO, result.getData());

        // Verify repository method call
        verify(triajeRepository).findById(1L);
        verify(genericMapper).mapTriajeEntityToTriajeDTO(triajeEntity);
    }

    @Test
    void buscarTriajeDtoOut_triageNotExist_returnsResponseBase() {
        // Arrange
        when(triajeRepository.findById(1L)).thenReturn(Optional.empty());

        // Act
        ResponseBase result = triajeAdapter.BuscarTriajeDtoOut(1L);

        // Assert
        assertEquals(404, result.getCode());
        assertEquals("Registro no encontrado", result.getMessage());
        assertEquals(null, result.getData());

        // Verify repository method call
        verify(triajeRepository).findById(1L);
        verifyNoInteractions(genericMapper); // No debería haber llamadas al mapper si no se encuentra el triaje
    }

    @Test
    void buscarAllEnableTriajeDtoOut_someTriajeEntitiesExist_returnsResponseBase() {
        // Arrange
        TriajeEntity triajeEntity1 = new TriajeEntity();
        triajeEntity1.setIdTriaje(1L);
        TriajeEntity triajeEntity2 = new TriajeEntity();
        triajeEntity2.setIdTriaje(2L);

        when(triajeRepository.findByEstado(true)).thenReturn(Arrays.asList(triajeEntity1, triajeEntity2));

        TriajeDTO triajeDTO1 = new TriajeDTO();
        triajeDTO1.setIdTriaje(1L);
        TriajeDTO triajeDTO2 = new TriajeDTO();
        triajeDTO2.setIdTriaje(2L);

        when(genericMapper.mapTriajeEntityToTriajeDTO(triajeEntity1)).thenReturn(triajeDTO1);
        when(genericMapper.mapTriajeEntityToTriajeDTO(triajeEntity2)).thenReturn(triajeDTO2);

        // Act
        ResponseBase result = triajeAdapter.BuscarAllEnableTriajeDtoOut();

        // Assert
        assertEquals(200, result.getCode());
        assertEquals("Solicitud exitosa", result.getMessage());

        Set<TriajeDTO> expectedDtoSet = new HashSet<>(Arrays.asList(triajeDTO1, triajeDTO2));
        assertEquals(expectedDtoSet, result.getData());

        // Verify repository method call
        verify(triajeRepository).findByEstado(true);
    }

    @Test
    void buscarAllEnableTriajeDtoOut_noTriajeEntitiesExist_returnsResponseBase() {
        // Arrange
        when(triajeRepository.findByEstado(true)).thenReturn(Arrays.asList());

        // Act
        ResponseBase result = triajeAdapter.BuscarAllEnableTriajeDtoOut();

        // Assert
        assertEquals(200, result.getCode());
        assertEquals("Solicitud exitosa", result.getMessage());
        assertEquals(new HashSet<>(), result.getData());

        // Verify repository method call
        verify(triajeRepository).findByEstado(true);
    }

    @Test
    void updateTriajeOut_consultaMedicaNotExist_returnsResponseBase() {
        // Arrange
        RequestTriaje requestTriaje = new RequestTriaje();
        requestTriaje.setIdConsulta(1L);

        when(consultaMedicaRepository.findById(1L)).thenReturn(Optional.empty());

        // Act
        ResponseBase result = triajeAdapter.UpdateTriajeOut(requestTriaje, "username");

        // Assert
        assertEquals(404, result.getCode());
        assertEquals("Consulta no encontrada", result.getMessage());
        assertEquals(null, result.getData());

        // Verify repository method call
        verify(consultaMedicaRepository).findById(1L);
        verifyNoInteractions(triajeRepository);
    }

    @Test
    void updateTriajeOut_consultaMedicaExistButInactive_returnsResponseBase() {
        // Arrange
        RequestTriaje requestTriaje = new RequestTriaje();
        requestTriaje.setIdConsulta(1L);
        ConsultaMedicaEntity consultaMedicaEntity = new ConsultaMedicaEntity();
        consultaMedicaEntity.setEstado(false);

        when(consultaMedicaRepository.findById(1L)).thenReturn(Optional.of(consultaMedicaEntity));

        // Act
        ResponseBase result = triajeAdapter.UpdateTriajeOut(requestTriaje, "username");

        // Assert
        assertEquals(404, result.getCode());
        assertEquals("Consulta no encontrada", result.getMessage());
        assertEquals(null, result.getData());

        // Verify repository method call
        verify(consultaMedicaRepository).findById(1L);
        verifyNoInteractions(triajeRepository);
    }

    @Test
    void updateTriajeOut_consultaMedicaAndTriajeExistAndActive_returnsResponseBase() {

        String usuario = "user01";
        RequestTriaje requestTriaje = new RequestTriaje();
        requestTriaje.setIdConsulta(1L);
        ConsultaMedicaEntity consultaMedicaEntity = new ConsultaMedicaEntity();
        consultaMedicaEntity.setEstado(true);
        TriajeEntity triajeEntity = new TriajeEntity();
        triajeEntity.setIdTriaje(1L);
        triajeEntity.setEstado(true);
        Optional<ConsultaMedicaEntity> optional = Optional.of(consultaMedicaEntity);
        Optional<TriajeEntity> optionalTriaje = Optional.of(triajeEntity);
        TriajeEntity triajeGuardar = new TriajeEntity();

        consultaMedicaEntity.setTriaje(triajeEntity);

        when(consultaMedicaRepository.findById(consultaMedicaEntity.getIdConsulta())).thenReturn(optional);
        when(triajeRepository.findById(optional.get().getTriaje().getIdTriaje())).thenReturn(optionalTriaje);
        //when(triajeRepository.save(optionalTriaje.get())).thenReturn(triajeGuardar);

        assertDoesNotThrow(() -> triajeAdapter.UpdateTriajeOut(requestTriaje, usuario));
        /*
        // Act
        ResponseBase result = triajeAdapter.UpdateTriajeOut(requestTriaje, usuario);

        // Assert
        assertEquals(200, result.getCode());
        assertEquals("Registro actualizado", result.getMessage());
        assertEquals(triajeGuardar, result.getData());

        // Verify repository method calls
        verify(consultaMedicaRepository).findById(consultaMedicaEntity.getIdConsulta());
        verify(triajeRepository).findById(triajeEntity.getIdTriaje());
        verify(triajeRepository).save(triajeEntity);*/
    }

    @Test
    void deleteTriajeOut_triageExist_returnsResponseBase() {
        // Arrange
        TriajeEntity triajeEntity = new TriajeEntity();
        triajeEntity.setIdTriaje(1L);

        when(triajeRepository.findById(1L)).thenReturn(Optional.of(triajeEntity));
        when(triajeRepository.save(triajeEntity)).thenReturn(triajeEntity);

        // Act
        ResponseBase result = triajeAdapter.DeleteTriajeOut(1L, "username");

        // Assert
        assertEquals(200, result.getCode());
        assertEquals("Registro eliminado", result.getMessage());
        assertEquals(triajeEntity, result.getData());

        // Verify repository method calls
        verify(triajeRepository).findById(1L);
        verify(triajeRepository).save(triajeEntity);
    }

    @Test
    void deleteTriajeOut_triageNotExist_returnsResponseBase() {
        // Arrange
        when(triajeRepository.findById(1L)).thenReturn(Optional.empty());

        // Act
        ResponseBase result = triajeAdapter.DeleteTriajeOut(1L, "username");

        // Assert
        assertEquals(404, result.getCode());
        assertEquals("Registro no encontrado", result.getMessage());
        assertEquals(null, result.getData());

        // Verify repository method call
        verify(triajeRepository).findById(1L);
        // Ensure that triajeRepository.save() is not called since triajeEntity doesn't exist
        verify(triajeRepository, never()).save(any());
    }
}